+++ doc/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.0.1] - 2023-10-25
+ ### `rna` and `dynaseq`
+ - Add a metric in HTML report - Corrected Barcodes.
+ - `celescope rna cells` add a new argument `--min_gene` to filter cells with low number of genes.
+
 ## [2.0.0] - 2023-10-25
  ### `rna` and `dynaseq`
  - Use Starsolo to quantify the gene expression.